requirements.txt
# Dependencias principales
torch>=2.0.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
safetensors>=0.3.1

# Procesamiento de imágenes
Pillow>=10.0.0

# Interfaz gráfica (viene con Python)
# tkinter (ya incluido en Python estándar)

# Utilidades
numpy>=1.24.0
